Fix use-after-free issue on array filters
This patch fix a use-after-free bug in the array filter that happens
when the filter's private struct (`pvt`) is freed while there are still
pending events referencing it. Specifically, this leads to the
destruction of the `arrayFreeList` and its mutex, which is later
accessed during field log cleanup.

When the queue thread eventually processes one of these `db_field_log`
entries and attempts to free the array via `freeListFree()`,
it tries to lock a now-destroyed mutex. This results in
`epicsMutexMustLock`
failing with `EINVAL`, triggering an assertion and suspending the queue
thread. The error looks like the following:

A call to 'assert(status == epicsMutexLockOK)'
    by thread 'PDB-event' failed in ../freeList/freeListLib.c line 169.
errlog: lost 1 messages
[    0x7f82695e7fb3]errlog: lost 3 messages
[    0x7f82695e10ce]errlog: lost 3 messages
[    0x7f82695cfb9a]errlog: lost 3 messages
[    0x7f826965c6e9]errlog: lost 3 messages
[    0x7f826965cce3]errlog: lost 3 messages
[    0x7f82695e2315]errlog: lost 3 messages
[    0x7f82691f7a92]errlog: lost 3 messages
[    0x7f82692797d0]errlog: lost 3 messages
EPICS Release EPICS R7.0.8.1-E3-7.0.8.1-patch.
Local time is 2025-03-20 08:53:04.260698231 CET
errlog: lost 2 messages
Thread PDB-event (0x558fad15ca90) suspended
callbackRequest: ERROR cbLow ring buffer full

To reproduce this data race for PVA was used a simple python script
to rapidly open and close monitor on array PVs with a random filter.
The same script cannot reproduce this on PVXS.

The fix is using a simple reference counter to ensure the pvt struct is
never destroyed if there are still references for arrays on the
arrayFreeList.